# [Performance] Improve import time with many (text) props by controlling concurrency better