--- a/adapters/repos/db/shard_write_batch_objects.go
+++ b/adapters/repos/db/shard_write_batch_objects.go
@@ -151,10 +151,23 @@ func (ob *objectsBatcher) storeSingleBatchInLSM(ctx context.Context,
 	}
 
 	wg := &sync.WaitGroup{}
+	concurrencyLimit := make(chan struct{}, _NUMCPU)
+
 	for j, object := range batch {
 		wg.Add(1)
 		go func(index int, object *storobj.Object) {
 			defer wg.Done()
+
+			// Acquire a semaphore to control the concurrency. Otherwise we would
+			// spawn one routine per object here. With very large batch sizes (e.g.
+			// 1000 or 10000+), this isn't helpuful and just leads to more lock
+			// contention down the line – especially when there's lots of text to be
+			// indexed in the inverted index.
+			concurrencyLimit <- struct{}{}
+			defer func() {
+				// Release the semaphore when the goroutine is done.
+				<-concurrencyLimit
+			}()
 
 			if err := ob.storeObjectOfBatchInLSM(ctx, index, object); err != nil {
 				errLock.Lock()