### What's being changed:

In the inverted index, all objects of a batch were processed in parallel. This unbounded concurrency led to unnecessary lock contention. Controlling the concurrency reduces contention and, therefore, improves import time.

## Validation - Sphere 1M (lots of text props)

Sphere 1M, new Python Client, all props default settings (auto-schema), batch size 1000

### Async importing

| build | importing | indexing | total |
| --- | --- | --- | --- |
| v1.22 (control) | 480s | 218s | 698s |
| this branch | 350s | 230s | 580s |

-> 35% faster importing
-> 19% faster total (importing+indexing)

### Sync importing

| build | total |
| --- | --- | 
| v1.22 (control) | 715s | 
| this branch | 630s |

-> 13% faster 

## Validation - SIFT1M

This is the control to make sure that this patch does not degrade performance for cases where there aren't many props. 

### Async importing

| build | importing | indexing | total |
| --- | --- | --- | --- |
| v1.22 (control) | 93s | 127s | 220s |
| this branch | 93s | 120s | 213s |

-> this branch is slightly faster. This seems to be within the margin of error, but at least we can see it's not any slower


Sync 

### Sync importing

| build | total |
| --- | --- | 
| v1.22 (control) | 241s | 
| this branch | 235s |

-> this branch is slightly faster. This seems to be within the margin of error, but at least we can see it's not any slower


### Review checklist

- [ ] Documentation has been updated, if necessary. Link to changed documentation:
- [x] Chaos pipeline run or not necessary. Link to pipeline: (running)
- [ ] All new code is covered by tests where it is reasonable.
- [ ] Performance tests have been run or not necessary.