@etiennedi 2023-10-29
Please wait with merging. It seems this might degrade performance if there are very few properties that need to be indexed.