@etiennedi 2023-10-30
I've added one more test scenario which is the "regular" case that this PR doesn't optimize for. The test makes sure that we don't see a regression on those cases. It is now ready to be merged from my side.